Conciliation de "tout effacer" avec "affiner votre situation"
Au passage, externatilsation de <Conversation/> de <Simulateur/>
